// src/app/models/ClientModel.ts
export interface ClientModel {
 
    code?: string;
    nom?:string;
    prenom?: string;
    adresse?: string;
    ville?: string;
    dateNaissance?: string;
    sexe?: string;
    situationFamiliale?: string;
    nombreEnfant?: string;
    numeroTelephone?: string;
    numeroPermis?: string;
    datePermis?: string;
    lieuPermis?: string;
}